requirements.txt
# packaging
pip>=9
setuptools>=26
twine==3.6.0
wheel>=0.29

# data science
numpy==1.18.5
pandas==1.3.3
pandas-profiling==3.1.0
scikit-image==0.18.3
scikit-learn==0.22

# tests/linter
black==21.10b0
coverage==6.1.2
flake8==3.9.2
pytest==6.2.5
yapf==0.31.0

# API
gcsfs==2021.11.0
google-cloud-storage==1.42.3
mlflow==1.21.0
s3fs==2021.11.0

# utilities
six>=1.14
joblib==1.0.1
memoized-property==1.0.3
termcolor==1.1.0
